Crate kicad_parse_gen

Kicad file format parser and generator library

Reexports

pub use error::*;

Modules

error
Kicad error handling code and types

footprint
Kicad footprint format handling

layout
Kicad layout format handling

project
Kicad project format handling

schematic
Kicad schematic format handling

symbol_lib
Kicad symbol library format handling

Enums

Expected
types of Kicad files that we expect to read

KicadFile
types of Kicad files that can be found

Sexp
a symbolic-expression Can be a string or a list or nothing

Traits

BoundingBox
calculate the bounding box of an item

FromSexp
convert from a symbolic-expression to something

Functions

from_sexp
convert from a symbolic-expression to something (dispatcher)

read_file
read a file

read_kicad_file
try to read a file, trying to parse it as the different formats and matching it against the Expected type

read_layout
read a file, expecting it to be a Kicad layout file

read_module
read a file, expecting it to be a Kicad module file

read_project
read a file, expecting it to be a Kicad project file

read_schematic
read a file, expecting it to be a Kicad schematic

read_symbol_lib
read a file, expecting it to be a Kicad symbol library file

write_file
write a file

write_layout
write out a kicad Layout to a file