Module kicad_parse_gen::footprint

Kicad footprint format handling

Structs

At
a location and rotation in a layout

Drill
a drill

Effects
text effects

Font
font attributes for text

FpArc
an arc

FpCircle
a circle

FpLine
a line

FpPoly
a polygon

FpText
text element

Layer
a pcb layer, with a side and a type

Layers
a list of layers

Model
a 3D model

Module
a Kicad module, with a name and a list of elements

Net
a net

Pad
a pad

Pts
a list of X-Y coordinates

Xy
X-Y element

Xyz
a 3D X-Y-Z coordinate

Enums

Element
elements that can be found in a Module

Justify
text justification

LayerSide
side of a layer

LayerType
type of a layer

PadShape
shape of a pad

PadType
type of a Pad

XyType
the type of X-Y element

Functions

module_to_string
convert a Kicad Module (footprint) to a String

parse
parse a &str to a Kicad Module