Enum kicad_parse_gen::schematic::FieldUpdate

pub enum FieldUpdate {
    New,
    Update(String),
    Same,
}

an indication of how a component field was updated

Variants
New

the field is new

Update(String)

the field is updated from specified old value to the new one

Same

the field already existed and is not updated